pyproject.toml
[tool.poetry]
name = "pingojo"
version = "0.1.0"
description = ""
authors = ["Your Name"]

[tool.poetry.dependencies]
python = "^3.11.4"
gunicorn = "^20.1.0"
django-allauth = "^0.58.2"
django-hashid-field = "^3.3.7"
beautifulsoup4 = "^4.11.2"
dj-database-url = "^2.1.0"
psycopg2 = "^2.9.5"
whitenoise = {extras = ["brotli"], version = "^6.4.0"}
pillow = "^10.0.1"
nltk = "^3.8.1"
pdfminer-six = "^20221105"
docx2txt = "^0.8"
sentry-sdk = "^1.25.1"
sendgrid = "^6.9.7"
djangorestframework = "^3.14.0"
pypdf2 = "^3.0.1"
django-cors-headers = "^3.14.0"
django = "4.2.7"
django-debug-toolbar = "^4.1.0"
selenium = "^4.15.2"
webdriver-manager = "^3.8.6"
openai = "^0.27.8"
html2text = "^2020.1.16"
django-simple-captcha = "^0.5.20"
djlint = "^1.34.0"

[tool.poetry.dev-dependencies]

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
